refactor functionality
