+++ src/api.v2/controllers/accessController.js
@@ -1,4 +1,4 @@
-const UserAccess = require('../model/UserAccess');
+const getUserRoles = require('../helpers/access/getUserRoles');
 const getLogger = require('../../utils/getLogger');
 
 const logger = getLogger('[AccessController] - ');
@@ -7,11 +7,11 @@ const getExperimentUsers = async (req, res) => {
   const { experimentId } = req.params;
 
   logger.log(`Fetching users for experiment ${experimentId}`);
-  const data = await new UserAccess().getExperimentUsers(experimentId);
+  const users = await getUserRoles(experimentId);
 
   logger.log(`Users fetched for experiment ${experimentId}`);
 
-  res.json(data);
+  res.json(users);
 };
 
 module.exports = {
